feat: add gravity tags component, not-found page, and update hero section
